requirements.txt
Flask>=2.2.0
requests>=2.28.0
python-dotenv>=0.21.0
flask-limiter>=3.5,<4
gunicorn>=20.1.0
redis>=5

